src/post-training/pyproject.toml
[project]
name = "gemini-grpo"
version = "0.1.0"
description = "GRPO fine-tuning pipeline where Gemini teaches Gemma using dialogue traces"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.2.0",
    "transformers>=4.46.0",
    "trl>=0.17.0",
    "datasets>=2.19.0",
    "accelerate>=0.33.0",
    "peft>=0.13.0",
    "google-genai>=1.0.0",
    "tqdm>=4.66.0",
    "dotenv>=0.9.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# Explicitly list the Python modules to avoid auto-discovery issues
py-modules = [
    "dataset_sgd_bandit",
    "env_compiler", 
    "gemini_judge",
    "jsonl_utils",
    "prompts",
    "reward_fn",
    "sgd_trace_parser",
]

[tool.uv]
# UV-specific configuration
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0", 
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py39"
